Update Docker tag and enhance file check and logging
The docker-build.yml file was amended to update the version tag of the krande/quetz-server from v10.5 to v10.6 for tracking and management purposes. The quetz/tasks/reindexing.py was modified to also consider files ending with '.conda' during the package filtering process, accommodating for the additional format. Added an additional debug log message in quetz/main.py to indicate successful file upload with the file's corresponding checksum to improve traceability.